+++ nmrs/src/core/mod.rs
@@ -11,6 +11,7 @@ pub(crate) mod connection_settings;
 pub(crate) mod connectivity;
 pub(crate) mod custom_connection;
 pub(crate) mod device;
+pub(crate) mod dns;
 pub(crate) mod ovpn_parser;
 pub(crate) mod rfkill;
 pub(crate) mod saved_connection;